PSD#:217339:Getting 900 Validation in Excess Amount Transfer When Amount Given Amount Decimals or Minor Currency.

+++ PFSWeb/src/main/java/com/pennant/pff/excess/web/ExcessTransferDialogCtrl.java
@@ -604,7 +604,7 @@ public class ExcessTransferDialogCtrl extends GFCBaseCtrl<FinExcessTransfer> {
 				if (StringUtils.isEmpty(this.excessReference.getDescription())) {
 					this.excessReference.setDescription(String.valueOf(BigDecimal.ZERO));
 				}
-				finExcessTransfer.setTransferToId(Long.parseLong(this.excessReference.getDescription()));
+				finExcessTransfer.setTransferToId((long) Double.parseDouble(this.excessReference.getDescription()));
 			}
 		} catch (WrongValueException we) {
 			wve.add(we);
